--- src/app/(auth)/logout/page.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Logout = () => <h2>Log out</h2>;
-
-export default Logout;
